+++ lib/src/who_called_me.dart
@@ -1,24 +1,25 @@
 class PackageCaller extends Caller {
-  PackageCaller(
+  PackageCaller._(
       {required super.stackPos,
       required super.functionName,
       required super.package,
       required super.fileName,
       required super.lineNumber,
-      required super.columnNumber});
+      required super.columnNumber})
+      : super._();
 
   @override
   String get path => "package:$package/$fileName:$lineNumber:$columnNumber";
 }
 
 class FileCaller extends Caller {
-  FileCaller(
+  FileCaller._(
       {required super.stackPos,
       required super.functionName,
       required super.fileName,
       required super.lineNumber,
       required super.columnNumber})
-      : super(package: null);
+      : super._(package: null);
 
   @override
   String get path => "file://$fileName $lineNumber:$columnNumber";
@@ -47,7 +48,7 @@ abstract class Caller {
   /// A path to where the function got called, either a `file:` or a `package:` path.
   String get path;
 
-  Caller({
+  Caller._({
     required this.stackPos,
     required this.functionName,
     required this.package,
@@ -59,18 +60,18 @@ abstract class Caller {
   @override
   String toString() => "Caller( $functionName $path )";
 
-  static Caller? parseCaller(String s) {
+  static Caller? _parseCaller(String s) {
     var packagePattern = RegExp(
         r'^#(?<index>\d+) +(?<fnname>.+) +\((package:)(?<package>[^/]+)/(?<filnamn>.*):(?<ln>\d+):(?<cn>\d+)\)*');
     var filePattern = RegExp(
         r'^#(?<index>\d+) +(?<fnname>.+) +\(file://(?<filnamn>.*):(?<ln>\d+):(?<cn>\d+)\)*');
-    var suspensionPatttern = RegExp(r'^<asynchronous suspension>');
+    var suspensionPattern = RegExp(r'^<asynchronous suspension>');
 
     var matches = packagePattern.allMatches(s);
 
     try {
       var match = matches.elementAt(0);
-      return PackageCaller(
+      return PackageCaller._(
           stackPos: int.parse(match.namedGroup("index")!),
           functionName: match.namedGroup("fnname")!,
           package: match.namedGroup("package")!,
@@ -81,7 +82,7 @@ abstract class Caller {
       try {
         var matches = filePattern.allMatches(s);
         var match = matches.elementAt(0);
-        return FileCaller(
+        return FileCaller._(
             stackPos: int.parse(match.namedGroup("index")!),
             functionName: match.namedGroup("fnname")!,
             fileName: match.namedGroup("filnamn")!,
@@ -89,11 +90,13 @@ abstract class Caller {
             columnNumber: int.parse(match.namedGroup("cn")!));
       } catch (e) {
         try {
-          var matches = suspensionPatttern.allMatches(s);
-          matches.elementAt(0);
+          // Ignoring lines with no info other than "<asynchronous suspension>"
+          // ignore: unused_local_variable
+          var matches = suspensionPattern.allMatches(s);
+          // print("Anonymous closure: ${matches.elementAt(0)} $s");
           return null;
         } catch (e) {
-          print("error parsing $s: $e");
+          // print("error parsing $s: $e");
           return null;
         }
       }
@@ -104,27 +107,30 @@ abstract class Caller {
   static Caller? get whoCalledWhoCalledMe {
     var frames = StackTrace.current.toString().split("\n");
     frames.removeLast();
-    return parseCaller(frames[3]);
+    return _parseCaller(frames[3]);
   }
 
   /// Returns info on the function calling the current function.
   static Caller? get whoCalledMe {
     var frames = StackTrace.current.toString().split("\n");
     frames.removeLast();
-    return parseCaller(frames[2]);
+    return _parseCaller(frames[2]);
   }
 
-  /// Returns info on the function calling the current function.
+  /// Returns info on the current function.
   static Caller? get whoAmI {
     var frames = StackTrace.current.toString().split("\n");
     frames.removeLast();
-    return parseCaller(frames[1]);
+    return _parseCaller(frames[1]);
   }
 
-  /// Returns info on the callers of this function
+  /// Returns info on each function in the full callstack of the current function.
   static List<Caller?> get callStack {
-    var frames = StackTrace.current.toString().split("\n");
-    frames.removeLast();
-    return [...frames.map(parseCaller)];
+    var frames = StackTrace.current.toString().split("\n")
+      // Skip parsing the last newline
+      ..removeLast()
+      // Skip parsing the call to callStack
+      ..removeAt(0);
+    return [...frames.map(_parseCaller)];
   }
 }