+++ CHANGELOG.md
@@ -2,3 +2,9 @@
 
 - Initial version as a standalone package.
 - Handles `package:` and `file:` stacktrace paths.
+
+# 0.0.2
+
+- Some documentation & lint fixes.
+- Made the Caller API cleaner.
+- The return value of Caller.callStack no longer starts with the call to itself.